requirements.txt
pandas==1.5.3
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.35.0
streamlit_option_menu==0.3.13
pytz==2024.1
pytest==8.2.2